+++ src/main/java/ltd/newbee/mall/newbeemall/dao/SkuCategoryMapper.java
@@ -5,6 +5,6 @@ import java.util.List;
 import ltd.newbee.mall.newbeemall.vo.SkuCategoryVo;
 
 public interface SkuCategoryMapper {
-	List<SkuCategoryVo> findGoodsInfoByGoodsCategoryId(Long categoryId,Integer pageNum, String orderBy,
+	List<SkuCategoryVo> findGoodsInfoByGoodsCategoryId(Integer categoryId,Integer pageNum, String orderBy,
 			String ascOrDesc);
 }
